refactor: Input UI를 사용하도록 수정
