+++ src/main/java/lotto/UI/Input.java
@@ -3,17 +3,17 @@ package lotto.UI;
 import camp.nextstep.edu.missionutils.Console;
 
 public class Input {
-    public String inputPurchasePrice() {
+    public static String inputPurchasePrice() {
         System.out.println("구입금액을 입력해 주세요.");
         return Console.readLine();
     }
 
-    public String inputWinningNumbers() {
+    public static String inputWinningNumbers() {
         System.out.println("당첨 번호를 입력해주세요.");
         return Console.readLine();
     }
 
-    public String inputBonusNumber() {
+    public static String inputBonusNumber() {
         System.out.println("보너스 번호를 입력해 주세요.");
         return Console.readLine();
     }
